Move InputContainer and TextContainer to Controls
