+++ src/components/sidebar/routePathView/routePathInfoTab/RoutePathInfoTab.tsx
@@ -13,8 +13,8 @@ import { IValidationResult } from '~/validation/FormValidator';
 import { IRoutePathPrimaryKey } from '~/models/IRoutePath';
 import RoutePathService from '~/services/routePathService';
 import QueryParams from '~/routing/queryParams';
-import InputContainer from '../../InputContainer';
-import TextContainer from '../../TextContainer';
+import InputContainer from '../../../controls/InputContainer';
+import TextContainer from '../../../controls/TextContainer';
 import LinkListView from './LinkListView';
 import { Button, Dropdown } from '../../../controls';
 import ButtonType from '../../../../enums/buttonType';
