+++ src/components/sidebar/routePathView/routePathListTab/RoutePathListNode.tsx
@@ -15,8 +15,8 @@ import { RoutePathStore } from '~/stores/routePathStore';
 import routePathLinkValidationModel from '~/models/validationModels/routePathLinkValidationModel';
 import NodeHelper from '~/util/nodeHelper';
 import navigator from '~/routing/navigator';
-import TextContainer from '../../TextContainer';
-import InputContainer from '../../InputContainer';
+import TextContainer from '../../../controls/TextContainer';
+import InputContainer from '../../../controls/InputContainer';
 import MultiTabTextarea from '../../linkView/MultiTabTextarea';
 import RoutePathListItem from './RoutePathListItem';
 import * as s from './routePathListItem.scss';
